Perfect! The refactoring is complete. Here's what I did:
## Changes Made

1. **Deleted** `frontend/src/hooks/usePendingApproval.ts` - the custom hook was unnecessary

2. **Updated** `TaskFollowUpSection.tsx`:
   - Removed import of `usePendingApproval`
   - Added import of `useEntries` from `@/contexts/EntriesContext`
   - Replaced the hook call with inline logic using `useEntries()` and `useMemo()`

## Result

The implementation is now:
- **Simpler** - one less file to maintain
- **More consistent** - follows the same pattern as `TodoPanel.tsx` which also uses `useEntries()` directly
- **Just as efficient** - still memoized, same performance
- **Easier to understand** - all the logic is in one place

The functionality remains exactly the same - the follow-up textbox (and all related buttons) will be disabled when there's a pending approval in the conversation entries. TypeScript type checking passes successfully.
